requirements.txt
absl-py==0.15.0
astunparse==1.6.3
bazel==0.0.0.20200723
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
flatbuffers==1.12
gast==0.3.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
keras-nightly==2.10.0.dev2022071907
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.15.0
tb-nightly==2.10.0a20220719
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tf-estimator-nightly==2.10.0.dev2022071908
tf-nightly==2.10.0.dev20220719
typing-extensions==3.7.4.3
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.1
